Support automatic line breaking at column limit
This feature is implemented as a hack,
as described by George Thomas in Issue #71.

Unfortunately, adding a column limit/max line
length right now is not idempotent. We mark
the feature with a warning accordingly in the README.

The actual logic changes of this commit are in:

* Ormolu/Printer/Combinators.hs
* Ormolu/Printer/Meat/Declaration/Value.hs

The rest is config changes, to surface this option.

Co-authored-by: CrystalSplitter <gamewhizzit@gmail.com>