Make derivedTyClsParensEq more rebust
The current implementation of this function relies on the fact that we
never output un-parenthesised deriving classes from formatting. This
means the right-hand argument is never a `DctSingle`. If it ever were,
the function would incorrectly return a `Difference` no matter what,
even if it were the exact same `DctSingle` as the left-hand argument.

Since we're about to start making those parenthesis conditional, this
won't work. Instead, we now use `considerEqualVia` and pattern match on
a comparison of single-to-singleton or singleton-to-single. I find this
a conceptually simpler approach anyway, and is more lenient in the way
we need.