Add trailing-section-operators
Fourmolu has a behavior (inherited from Ormolu) where so-called
"section" operators (those that are `infixr 0`, such as `$`) are left
trailing, with the continuation of expressions indented below, rather
than placed on the next line as most operators are.

It means code comes out like this:

    foo $
      bar $
        baz $
          bat $
            quix

Instead of this:

    foo
      $ bar
      $ baz
      $ bat
      $ quix

This behavior is not desired by some users for aesthetic reasons, and
can be even more confusing since it's unreliable: it requires fourmolu
can determine the fixity of `$` correctly, which depends on dependencies
in the cabal file and imports or re-export knowledge. All of which can
make the behavior change from place to place confusing users and can
generating differences between projects for seemingly know reason.

This new option allows disabling that behavior.

Closes #405.